server/pyproject.toml
[tool.poetry]
name = "lunarwebsite"
version = "0.1.0"
description = "The backend to the lunarian website"
authors = ["BadPythonCoder <xsare435@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.103.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
tortoise-orm = {extras = ["aiosqlite"], version = "^0.20.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
